socialnetwork/requirements.txt
asgiref==3.8.1
betterpath==0.2.2
click==8.1.7
Django==3.2
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
l==0.11.0
psycopg2-binary==2.8.6
PyJWT==2.8.0
pytz==2024.1
rest-framework-simplejwt==0.0.2
sqlparse==0.5.0
typing_extensions==4.12.2
vcversioner==2.16.0.0
zope.interface==6.4.post2
